+++ src/onegov.org/onegov/org/views/reservation.py
@@ -118,7 +118,12 @@ def reserve_allocation(self, request):
 
 @OrgApp.json(model=Reservation, request_method='DELETE', permission=Public)
 def delete_reservation(self, request):
-    request.assert_valid_csrf_token()
+
+    # anonymous users do not get a csrf token (it's bound to the identity)
+    # therefore we can't check for it -> this is not a problem since
+    # anonymous users do not really have much to lose here
+    if request.is_logged_in:
+        request.assert_valid_csrf_token()
 
     resource = request.app.libres_resources.by_reservation(self)